+++ Tests/OAuth/ResourceOwner/AmazonResourceOwnerTest.php
@@ -37,12 +37,10 @@ json;
                  'authorization_url' => 'https://www.amazon.com/ap/oa',
                  'access_token_url'  => 'https://api.amazon.com/auth/o2/token',
                  'infos_url'         => 'https://api.amazon.com/user/profile',
-
-                 'scope'             => 'profile',
             ),
             $options
         );
 
-        return new AmazonResourceOwner($this->buzzClient, $httpUtils, $options, $name);
+        return new AmazonResourceOwner($this->buzzClient, $httpUtils, $options, $name, $this->storage);
     }
 }