+++ Tests/OAuth/ResourceOwner/FacebookResourceOwnerTest.php
@@ -45,7 +45,7 @@ json;
         $resourceOwner = $this->createResourceOwner('facebook', array('display' => 'popup'));
         $this->assertEquals('popup', $resourceOwner->getOption('display'));
         $this->assertEquals(
-            $this->options['authorization_url'] . '&response_type=code&client_id=clientid&redirect_uri=http%3A%2F%2Fredirect.to%2F&display=popup',
+            $this->options['authorization_url'] . '&response_type=code&client_id=clientid&state=random&redirect_uri=http%3A%2F%2Fredirect.to%2F&display=popup',
             $resourceOwner->getAuthorizationUrl('http://redirect.to/')
         );
     }
@@ -75,6 +75,6 @@ json;
             $options
         );
 
-        return new FacebookResourceOwner($this->buzzClient, $httpUtils, $options, $name);
+        return new FacebookResourceOwner($this->buzzClient, $httpUtils, $options, $name, $this->storage);
     }
 }